pyproject.toml
[project]
name = "tutorapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.49.0",
    "pyside6>=6.10.0",
]
